KVM: arm64: Introduce a hyp buddy page allocator
The hyp code will require a basic form of memory management in order to
allocate and free memory pages at EL2. This is needed for various
use-cases, including (but not limited to) the creation of hyp mappings,
the allocation of stage 2 page tables, or the allocation of SPCI buffers.
Note however, that the EL2 users of a memory allocator can have strict
requirements.

For example, the creation of concatenated page tables requires to
allocate physically contiguous pages, with a given alignment. It is also
desirable for the memory allocator to support the usage of multiple and
independent memory heaps, to have independent memory 'contexts' for each
guest. Each guest heap will be used to allocate memory 'owned' by a
guest, such as its stage 2 page tables, or its SPCI buffers. The heaps
will also need to be extensible on demand, to avoid large memory
reservations upfront for each guest. If a guest heap runs out of free
pages, the host will be relied upon to donate pages to append to the
heap.

To address these use-case, introduce a simple memory allocator in the
hyp code. The allocator is designed as a conventional 'buddy allocator',
working with a page granularity. It allows to allocate and free
physically contiguous pages from memory 'pools', with a guaranteed order
alignment in the PA space. Each page in a memory pool is associated
with a struct hyp_page which holds the page's metadata, including its
refcount, as well as its current order, hence mimicking the kernel's
buddy system in the GFP infrastructure. The hyp_page metadata are made
accessible through a hyp_vmemmap, following the concept of
SPARSE_VMEMMAP in the kernel.

Signed-off-by: Quentin Perret <qperret@google.com>