KVM: arm64: Introduce hyp memory management setup
At boot time, in the nVHE case, the hypervisor needs to be initialized
to be able to manage its own memory. As such, add that setup code to the
kvm/hyp folder.

Essentially, the hypervisor initialization is initiated by the host,
which donates a contiguous memory region to EL2. The current assumption
is for the host to create temporary s1 hyp mappings for __hyp_text and
friends, to simplify the bootstrap procedure at EL2.

During the init hcall, the hypervisor uses the trivial early page
allocator to re-create page-tables mapping __hyp_text, the hypervisor
memory region, the __hyp_bss, the portion of hyp_vmemmap corresponding
to the hyp memory region, among other things. It then jumps back in the
idmap page, switches to use the newly-created pgd (instead of the
temporary one provided by the host) and then installs the full-fledged
buddy allocator which will then be the only one in used from then on.

Note that for the sake of symplifying the review, this only introduces
the code doing this operation, without actually plugging it in the
build system. This will be done in a subsequent patch, which will
introduce the necessary changes to the host.

Credits to Will for __kvm_init_switch_pgd.

Co-authored-by: Will Deacon <will@kernel.org>
Signed-off-by: Quentin Perret <qperret@google.com>