KVM: arm64: Separate SError detection from handling
When exiting a guest, just check whether there is an SError pending and
set the bit in the exit code. The hyp fixup then handles it, if
required.

With this separation, the assembly stays simple and leaves flexibility
in deciding whether or not to handle the SError. For example, were the
host to ever be viewed as a vcpu, the SErrors should be left pending for
the host to hanle itself rather than converting them to virtual SErrors.

Signed-off-by: Andrew Scull <ascull@google.com>