KVM: arm64: The ubsan handlers are defined inside EL2.
The handlers are implemented inside EL2, as empty functions for the
moment, so the undefined behaviours that are triggered at EL2 will be
linked to them, not to the one that lives at EL1.
In this way, enabling UBSAN_MISC won't cause a Kernel Panic if an undefined
behaviour is detected inside EL2.