KVM: arm64: Add support for creating and checking a buffer
The KVM_ARM_DEBUG_BUFFER option enables checking the buffer every time when
the state returns back to EL1 after an hvc call.
The __kvm_arm_check_debug_buffer checks if there is any data from EL2 UBSAN,
or the moment, but it can be extended into checking data from other tools
like KCOV and GCOV.
The macros can be used to define a buffer at EL2, and to declare it at EL1.
The kvm_debug_buffer_ind is an EL1 function that can return an interval
[Left, Right] and the number of used slots. The idea is to iterate from
[Right, BUFF_Len] and from [0, Left] as long as there are written slots
available.

Signed-off-by: George Popescu <georgepope@google.com>