KVM: arm64: Add a buffer that can pass UBSAN data from EL2 to EL1.
The data, which is collected from UBSAN handlers that lives at EL2, is
stored into this circular buffer.
This buffer stores only UBSAN data because it should not be preoccupied
by other mechanisms data structures and functionalities.
Also, for the moment the buffer is mapped inside .bss, where both EL1 and EL2
have Read/Write rights, but in the future this will change and EL1 will not
be able to acess EL2's .bss. At that point the buffer will only need to be
mapped in order for this to work.

Signed-off-by: George Popescu <georgepope@google.com>