KVM: arm64: Add support for logging in KCOV
KCOV shares it's own buffer with userspace. The KCOV buffer is attached
to a thread for which coverage is enabled. Unfortunately that
buffer can not be shared with the hyp/nVHE code.

Define a kvm_debug_buffer to log messages that are coming from the
kcov handlers. The buffer is responsible to pass data from hyp/nVHE
to the kernel and it's mapped per_cpu when the hypervisor is
initialized.

The buffer is checked whenever the state returns from a hyp_call.
That is the moment where the data from hyp/nVHE should be passed
to the KCOV kernel buffer.

Signed-off-by: George Popescu <georgepope@android.com>