KVM: arm64: Add support for KCOV_ENABLE_COMPARISON for hyp/nVHE
Collecting comparison operands is similar to coverage collection.
Store all the comparison arguments, instead of storing only the
instruction pointer.

KCOV stores a part of the arguments in the cases array, which
is stored on the .data region. Map the data region for hyp to be
able to read from it.

Signed-off-by: George Popescu <georgepope@android.com>